Lack Syntax
I have a working SimpleToDo Android app for CodePath admissions
submittal, and am attempting to add the cascading edit text and button.

Unfortunately I can find no usable example of syntax from which to
assemble a working set of code, the paragraphs in the prework are clear
enough, and the comments in the cliffnotes make block diagram sense,
but the syntax is illusive.

Is there such a set of examples please?  This may never be soon enough
for CodePath subbmission, but I would like to learn to cascade an
activity in Java, if indeed it can be done in Java.
:-)